refactor(web): convert internal prediction methods to stateless format
